fix: move the get order util to util Description: Move the get order util to util folder SONIC
