+++ commerce_coordinator/apps/commercetools/tests/raw_ct_order.json
@@ -41,7 +41,7 @@
           },
           {
             "name": "mode",
-            "value": "certified"
+            "value": "professional"
           },
           {
             "name": "date-created",
